// old bot (typescript)/src/DestinyAPI/Types/Destiny2/DestinyDatabaseTable.ts
enum DestinyDatabaseTable {
  DestinyPlaceDefinition = "DestinyPlaceDefinition",
  DestinyActivityDefinition = "DestinyActivityDefinition",
  DestinyActivityTypeDefinition = "DestinyActivityTypeDefinition",
  DestinyClassDefinition = "DestinyClassDefinition",
  DestinyGenderDefinition = "DestinyGenderDefinition",
  DestinyInventoryBucketDefinition = "DestinyInventoryBucketDefinition",
  DestinyRaceDefinition = "DestinyRaceDefinition",
  DestinyTalentGridDefinition = "DestinyTalentGridDefinition",
  DestinyUnlockDefinition = "DestinyUnlockDefinition",
  DestinySandboxPerkDefinition = "DestinySandboxPerkDefinition",
  DestinyStatGroupDefinition = "DestinyStatGroupDefinition",
  DestinyFactionDefinition = "DestinyFactionDefinition",
  DestinyVendorGroupDefinition = "DestinyVendorGroupDefinition",
  DestinyRewardSourceDefinition = "DestinyRewardSourceDefinition",
  DestinyItemCategoryDefinition = "DestinyItemCategoryDefinition",
  DestinyDamageTypeDefinition = "DestinyDamageTypeDefinition",
  DestinyActivityModeDefinition = "DestinyActivityModeDefinition",
  DestinyMedalTierDefinition = "DestinyMedalTierDefinition",
  DestinyAchievementDefinition = "DestinyAchievementDefinition",
  DestinyActivityGraphDefinition = "DestinyActivityGraphDefinition",
  DestinyBondDefinition = "DestinyBondDefinition",
  DestinyCollectibleDefinition = "DestinyCollectibleDefinition",
  DestinyDestinationDefinition = "DestinyDestinationDefinition",
  DestinyEquipmentSlotDefinition = "DestinyEquipmentSlotDefinition",
  DestinyEventCardDefinition = "DestinyEventCardDefinition",
  DestinyStatDefinition = "DestinyStatDefinition",
  DestinyInventoryItemDefinition = "DestinyInventoryItemDefinition",
  DestinyItemTierTypeDefinition = "DestinyItemTierTypeDefinition",
  DestinyLocationDefinition = "DestinyLocationDefinition",
  DestinyLoreDefinition = "DestinyLoreDefinition",
  DestinyMaterialRequirementSetDefinition = "DestinyMaterialRequirementSetDefinition",
  DestinyMetricDefinition = "DestinyMetricDefinition",
  DestinyObjectiveDefinition = "DestinyObjectiveDefinition",
  DestinyPlugSetDefinition = "DestinyPlugSetDefinition",
  DestinyPowerCapDefinition = "DestinyPowerCapDefinition",
  DestinyPresentationNodeDefinition = "DestinyPresentationNodeDefinition",
  DestinyProgressionDefinition = "DestinyProgressionDefinition",
  DestinyProgressionLevelRequirementDefinition = "DestinyProgressionLevelRequirementDefinition",
  DestinyRecordDefinition = "DestinyRecordDefinition",
  DestinySackRewardItemListDefinition = "DestinySackRewardItemListDefinition",
  DestinySandboxPatternDefinition = "DestinySandboxPatternDefinition",
  DestinySeasonDefinition = "DestinySeasonDefinition",
  DestinySeasonPassDefinition = "DestinySeasonPassDefinition",
  DestinySocketCategoryDefinition = "DestinySocketCategoryDefinition",
  DestinySocketTypeDefinition = "DestinySocketTypeDefinition",
  DestinyTraitDefinition = "DestinyTraitDefinition",
  DestinyTraitCategoryDefinition = "DestinyTraitCategoryDefinition",
  DestinyVendorDefinition = "DestinyVendorDefinition",
  DestinyMilestoneDefinition = "DestinyMilestoneDefinition",
  DestinyActivityModifierDefinition = "DestinyActivityModifierDefinition",
  DestinyReportReasonCategoryDefinition = "DestinyReportReasonCategoryDefinition",
  DestinyArtifactDefinition = "DestinyArtifactDefinition",
  DestinyBreakerTypeDefinition = "DestinyBreakerTypeDefinition",
  DestinyChecklistDefinition = "DestinyChecklistDefinition",
  DestinyEnergyTypeDefinition = "DestinyEnergyTypeDefinition",
  DestinyHistoricalStatsDefinition = "DestinyHistoricalStatsDefinition",
}
export default DestinyDatabaseTable;