// old bot (typescript)/src/DestinyAPI/Types/Destiny2/interfaces.ts
export interface DestinyIconSequenceDefinition {
  readonly frames: string[];
}
// ...
export interface DestinyDisplayPropertiesDefinition {
  readonly description: string;
  readonly name: string;
  /**
   * Note that "icon" is sometimes misleading, and should be interpreted in the
   * context of the entity. For instance, in Destiny 1 the
   * DestinyRecordBookDefinition's icon was a big picture of a book.
   *
   * But usually, it will be a small square image that you can use as... well, an
   * icon.
   *
   * They are currently represented as 96px x 96px images.
   */
  readonly icon: string;
  readonly iconSequences: DestinyIconSequenceDefinition[];
  /**
   * If this item has a high-res icon (at least for now, many things won't), then the
   * path to that icon will be here.
   */
  readonly highResIcon: string;
  readonly hasIcon: boolean;
}
// ...
export interface DestinyPlugSetDefinition {
  /**
   * If you want to show these plugs in isolation, these are the display properties
   * for them.
   */
  readonly displayProperties: DestinyDisplayPropertiesDefinition;
  /**
   * This is a list of pre-determined plugs that can be plugged into this socket,
   * without the character having the plug in their inventory.
   *
   * If this list is populated, you will not be allowed to plug an arbitrary item in
   * the socket: you will only be able to choose from one of these reusable plugs.
   */
  readonly reusablePlugItems: DestinyItemSocketEntryPlugItemRandomizedDefinition[];
  /**
   * Mostly for our debugging or reporting bugs, BNet is making "fake" plug sets in a
   * desperate effort to reduce socket sizes.
   *
   * If this is true, the plug set was generated by BNet: if it looks wrong, that's
   * a good indicator that it's bungie.net that fucked this up.
   */
  readonly isFakePlugSet: boolean;
  /**
   * The unique identifier for this entity. Guaranteed to be unique for the type of
   * entity, but not globally.
   *
   * When entities refer to each other in Destiny content, it is this hash that they
   * are referring to.
   */
  readonly hash: number;
  /** The index of the entity as it was found in the investment tables. */
  readonly index: number;
  /**
   * If this is true, then there is an entity with this identifier/type combination,
   * but BNet is not yet allowed to show it. Sorry!
   */
  readonly redacted: boolean;
}
export interface DestinyItemSocketEntryPlugItemRandomizedDefinition {
  readonly craftingRequirements: DestinyPlugItemCraftingRequirements;
  /**
   * Indicates if the plug can be rolled on the current version of the item. For
   * example, older versions of weapons may have plug rolls that are no longer
   * possible on the current versions.
   */
  readonly currentlyCanRoll: boolean;
  /**
   * The hash identifier of a DestinyInventoryItemDefinition representing the plug
   * that can be inserted.
   *
   * Mapped to DestinyInventoryItemDefinition in the manifest.
   */
  readonly plugItemHash: number;
}
export interface DestinyPlugItemCraftingRequirements {
  readonly unlockRequirements: DestinyPlugItemCraftingUnlockRequirement[];
  /** If the plug has a known level requirement, it'll be available here. */
  readonly requiredLevel?: number;
  /** Mapped to DestinyMaterialRequirementSetDefinition in the manifest. */
  readonly materialRequirementHashes: number[];
}